Fixed the code so that all tests would pass, some styles have been modified
